requirements.txt
python-telegram-bot>=22.7
httpx>=0.27
anyio>=4.0
pytest>=9.0
pytest-asyncio>=1.0
